crypto/ecies: fix build

+++ crypto/ecies/ecies.go
@@ -175,7 +175,7 @@ func concatKDF(hash hash.Hash, z, s1 []byte, kdLen int) (k []byte, err error) {
 	counter := []byte{0, 0, 0, 1}
 	k = make([]byte, 0)
 
-	for i := 0; i <= reps; i++ {
+	for i := int64(0); i <= reps; i++ {
 		hash.Write(counter)
 		hash.Write(z)
 		hash.Write(s1)
